refactor(Restaurant): 컴포넌트 .ts파일로 변경
ImgFileName 타입 지정, imgFileName 객체로 따로 분리
